requirements.txt
apturl==0.5.2
asgiref==3.5.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==4.0.2
django-bootstrap-form==3.4
django-crispy-forms==1.14.0
django-filter==21.1
django-floppyforms==1.9.0
django-heroku==0.3.1
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.6.0
fonttools==4.31.2
future==0.18.2
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
keyring==18.0.1
kiwisolver==1.4.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.5.1
monotonic==1.5
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.0
olefile==0.46
packaging==21.3
paramiko==2.6.0
pexpect==4.6.0
phonenumbers==8.12.46
Pillow==7.0.0
platformdirs==2.5.1
plotly==5.7.0
protobuf==3.6.1
psycopg2==2.8.4
psycopg2-binary==2.9.3
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.7
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-stdnum==1.17
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
systemd-python==234
tenacity==8.0.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.14.1
wadllib==1.3.3
whitenoise==6.0.0
xkit==0.0.0
